VIX-3909 Create Prop Wizard View Model Base Classes
VIX-3909 Create Prop Wizard View Model Base Classes
